+++ R/print.bc.mean.sd.R
@@ -9,7 +9,7 @@
 #' @examples
 #' ## Generate S2 summary data
 #' set.seed(1)
-#' n <- 1000
+#' n <- 100
 #' x <- stats::rlnorm(n, 2.5, 1)
 #' quants <- stats::quantile(x, probs = c(0.25, 0.5, 0.75))
 #' obs.mean <- mean(x)