+++ NEWS.md
@@ -9,6 +9,8 @@ Package Updates
 -   Added `plot.qe.fit()`, an S3 plot method for objects returned by
     `qe.fit()`. Consequently, added ‘graphics’ and ‘grDevices’ to
     Imports
+-   Added `summary.qe.mean.sd()`, an S3 summary method for objects
+    returned by `qe.mean.sd()`
 -   Removed prior on the power parameter lambda in `bc.mean.sd()`
 -   Added bounds on the sample size in `bc.mean.sd()` and `qe.fit()`
 -   Added URL and BugReport webpage to the DESCRIPTION file